refactor Authors component and AuthorPills
- created pills for 2 authors and 3+ authors;
- moved Author pills to a single file located in Authors/AuthorPills;
- fixed imports
