+++ src/app/_components/Authors/AuthorPills/styles.module.css
@@ -1,0 +1,46 @@
+/* General Styles */
+.authorImage {
+  position: relative;
+  width: 34px;
+  height: 34px;
+  border-radius: 50%;
+  transition: 0.5s ease;
+}
+
+.socials {
+  display: flex;
+  gap: 16px;
+}
+
+/* Author Pills */
+.authorPill,
+.twoAuthorPill {
+  display: flex;
+  gap: 10px;
+  align-items: center;
+  background-color: var(--soft-white-100);
+  width: fit-content;
+  font-weight: bold;
+  border-radius: 100px;
+  padding: 5px 15px 5px 5px;
+  color: var(--dark-rock-800);
+  font-size: var(--size-12);
+  font-family: var(--colfax);
+  transition: 0.5s ease;
+}
+
+.authorPill:hover,
+.twoAuthorPill:hover {
+  color: #C1C1C1;
+}
+
+.authorPill:hover .authorImage,
+.twoAuthorPill:hover .authorImage {
+  opacity: 0.5;
+}
+
+/* Multiple Authors */
+.authorTwo,
+.authorThree {
+  margin-left: -20px;
+}
