+++ src/app/_components/PostSummary/index.tsx
@@ -2,8 +2,9 @@ import { Blogpost } from "@/payload-types";
 import { estimateReadTime } from "../../_utilities/estimateReadTime";
 import { formatDateTime } from "../../_utilities/formatDateTime";
 import ArchiveButton from "../ArchiveButton";
-import AuthorPill from "../AuthorPill";
+
 import styles from "./styles.module.css";
+import Authors from '@/app/_components/Authors'
 
 export default function PostSummary({ post }) {
   const { title, publishedAt, content, authors } = post;
@@ -23,7 +24,7 @@ export default function PostSummary({ post }) {
       {/* Author info */}
       <div className={styles.authorInfo}>
         <p className={styles.outline}>WRITTEN BY</p>
-        <AuthorPill author={post.authors[0]} />
+        <Authors authors={authors} />
         Social Links
       </div>
     </div>
